rename Algorithm to Algorithms and add namespace
